+++ src/App.js
@@ -8,7 +8,7 @@ import {
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 import generateAcak from './libs/password-ga';
 import generateKata from './libs/password-ga-kata';
-import { demutatePassword } from './libs/mutation-common';
+import { demutatePassword } from './libs/common';
 
 import './App.css';
