+++ src/libs/common/array.js
@@ -15,3 +15,15 @@ export function pickNRandom(arr, n) {
 export function pickOneRandom(arr) {
   return arr[Math.floor(Math.random() * arr.length)];
 }
+
+export const getMinMax = (arr) => {
+  let min = arr[0];
+  let max = arr[0];
+  let i = arr.length;
+
+  while (i--) {
+    min = arr[i] < min ? arr[i] : min;
+    max = arr[i] > max ? arr[i] : max;
+  }
+  return { min, max };
+};
